Move specific web server utils to server module
Will allow to install cli and rpc without having to install web dependencies
